+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/Spec/Conway.hs
@@ -16,6 +16,7 @@ spec :: Spec
 spec = describe "Conway conformance tests" $ do
   xprop "UTXO" $ conformsToImpl @"UTXO" @ConwayFn @Conway
   prop "GOV" $ conformsToImpl @"GOV" @ConwayFn @Conway
+  prop "CERT" $ conformsToImpl @"CERT" @ConwayFn @Conway
   describe "Generators" $ do
     let
       genEnv = do